Moved game rendering to separate file
This is gonna get out of hand if it stays in game.c
