+++ OpenGLGame/OpenGLGame.vcxproj
@@ -144,6 +144,7 @@ xcopy "$(SolutionDir)OpenGLGame\assets" "$(OutDir)assets\" /S /Y</Command>
     <ClCompile Include="game.c" />
     <ClCompile Include="clock.c" />
     <ClCompile Include="game_load.c" />
+    <ClCompile Include="game_render.c" />
     <ClCompile Include="input.c" />
     <ClCompile Include="menu.c" />
     <ClCompile Include="random.c" />
